7.4.2d-e

+++ 7 module/7.4.2b.c
@@ -16,12 +16,12 @@ int main()
     char str[101], *p[n];
     for (int i = 0; i < n + 1; i++)
     {
-        if (fgets (str, 100, stdin) != NULL)
+        if (fgets(str, 100, stdin) != NULL)
         {
-            p[i] = malloc (strlen(str) + 1);
+            p[i] = malloc(strlen(str) + 1);
             if (p[i])
             {
-                strcpy (p[i], str);
+                strcpy(p[i], str);
             }
         }
     }
@@ -33,7 +33,7 @@ void func(char *p[])
     printf("\nCorrect lines:\n");
     for (int i = 0; p[i] != NULL; i++)
     {
-        if ((int)strlen(p[i]) > 3)
+        if (strlen(p[i]) > 3)
         {
             printf("%3s", p[i]);
         }
