+++ 7 module/7.4.2c.c
@@ -16,12 +16,12 @@ int main()
     char str[101], *p[n];
     for (int i = 0; i < n + 1; i++)
     {
-        if (fgets (str, 100, stdin) != NULL)
+        if (fgets(str, 100, stdin) != NULL)
         {
-            p[i] = malloc (strlen(str) + 1);
+            p[i] = malloc(strlen(str) + 1);
             if (p[i])
             {
-                strcpy (p[i], str);
+                strcpy(p[i], str);
             }
         }
     }
@@ -31,14 +31,14 @@ int main()
 void func(char *p[])
 {
     char *ptr = p[0];
-    int max = (int)strlen(p[0]) - 1;
+    int max = strlen(p[0]) - 1;
     printf("\nThe longest line is:\n");
     for (int i = 0; p[i] != NULL; i++)
     {
-        if ((int)(strlen(p[i]) - 1) > max)
+        if ((strlen(p[i]) - 1) > max)
         {
             ptr = p[i];
-            max = (int)strlen(p[i]) - 1;
+            max = strlen(p[i]) - 1;
         }
     }
     printf("%s", ptr);
